Move utils styling files into a styling folder
